requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
idna==3.3
newsapi-python==0.2.6
python-decouple==3.5
requests==2.27.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
